todo-list-app/src/api/todo_list/serializers/todo_list.py
from rest_framework import serializers
from api.todo_list.models import TodoList


class TodoListSerializer(serializers.ModelSerializer):

    class Meta:
        model = TodoList
        fields = "__all__"