todo-list-app/requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
Faker==13.11.1
idna==3.3
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
python-dotenv==1.0.0
django-cors-headers==3.13.0
psycopg2==2.9.3
